Switch windows to using native-reg to avoid permissions
